requirements.txt
streamlit>=1.37
requests>=2.32
geopy>=2.4
python-dotenv>=1.0
groq>=0.11.0
langgraph>=0.2.35